+++ reader.go
@@ -104,9 +104,10 @@ func (e UnknownDatabaseTypeError) Error() string {
 
 // Open takes a string path to a file and returns a Reader struct or an error.
 // The database file is opened using a memory map. Use the Close method on the
-// Reader object to return the resources to the system.
-func Open(file string) (*Reader, error) {
-	reader, err := maxminddb.Open(file)
+// Reader object to return the resources to the system. Any reader options
+// provided are passed directly to maxminddb.Open.
+func Open(file string, options ...maxminddb.ReaderOption) (*Reader, error) {
+	reader, err := maxminddb.Open(file, options...)
 	if err != nil {
 		return nil, err
 	}
@@ -117,9 +118,10 @@ func Open(file string) (*Reader, error) {
 // OpenBytes takes a byte slice corresponding to a GeoIP2/GeoLite2 database
 // file and returns a Reader struct or an error. Note that the byte slice is
 // used directly; any modification of it after opening the database will result
-// in errors while reading from the database.
-func OpenBytes(bytes []byte) (*Reader, error) {
-	reader, err := maxminddb.OpenBytes(bytes)
+// in errors while reading from the database. Any reader options provided are
+// passed directly to maxminddb.OpenBytes.
+func OpenBytes(bytes []byte, options ...maxminddb.ReaderOption) (*Reader, error) {
+	reader, err := maxminddb.OpenBytes(bytes, options...)
 	if err != nil {
 		return nil, err
 	}
@@ -134,8 +136,8 @@ func OpenBytes(bytes []byte) (*Reader, error) {
 //
 // Deprecated: Use OpenBytes instead. FromBytes will be removed in a future
 // version.
-func FromBytes(bytes []byte) (*Reader, error) {
-	return OpenBytes(bytes)
+func FromBytes(bytes []byte, options ...maxminddb.ReaderOption) (*Reader, error) {
+	return OpenBytes(bytes, options...)
 }
 
 func getDBType(reader *maxminddb.Reader) (databaseType, error) {